Pong Game Implemented.
Engine:
- Removed IRenderer.
* Key no longer imports Windows.h.
* Fixed compiler warning in Renderer.cpp.
* Locked game at 60FPS again.
+ Added forward declarations for windows structures in Window.h.

Demo:
- Removed Actor and subclasses.
+ Added Rect struct.
+ Added Object struct.
+ Added Camera struct.
- Removed Skybox shader.
+ Implemented Pong Mechanics.

+++ TinyEngine/Renderer.cpp
@@ -222,7 +222,7 @@ void TinyEngine::Renderer::DrawMesh(Mesh* mesh, std::vector<Material*> materials
 	auto* material = materials[0];
 	if (mesh->GetNumMeshParts() > 0)
 	{
-		for (int i = 0; i < mesh->GetNumMeshParts(); i++)
+		for (size_t i = 0; i < mesh->GetNumMeshParts(); i++)
 		{
 			auto part = mesh->GetMeshPart(i);
 			if (i < materials.size())
